[mojo-stdlib] Rewrite IntLiteral and FloatLiteral to use dependent types.
This patch rewrites the IntLiteral and FloatLiteral types to define away
bugs due to their long-standing but incorrect design. The problem is
that
the "value" that these types represent are infinite precision literals
that are
only representable at compile time, not materializable at runtime. The
previous design held this as runtime state, which led to bugs in the
corner
cases, e.g. this would crash the compiler because they can't be
codegen'd:

```
@export
fn test(a: IntLiteral) -> IntLiteral: return a+1
```

We can correct this by building on the recent improvement to dependent
type support (eg parameter inference bugfixes), the movement of all the
related intrinsics to an attribute form, and the
`always_inline(builtin)`
feature being settled, we can correct this by moving these to being
parameters on the literal types instead of runtime values. In short:

```
struct IntLiteral:
    var state : __mlir_type.`!kgen.int_literal`
struct IntLiteral[value: __mlir_type.`!kgen.int_literal`]:
   # .. no state
```

In addition to defining away a ton of compiler crashes, this also
greatly
simplifies the internal of the compiler by removing all the MLIR
operations
that were needed to support the old design.

This fixes MOCO-1628, and provides a pattern that we can bring forward
to `StringLiteral` as well.

QUANTIZATION_ORIG_REV_ID: cc0f00802fab0592d1ae788805cbe3ce89f57e36