[Nanobind][Stubgen] Generate stubfiles from nanobind extensions (#54251)
Motivation: These files can easily get out of sync with the C++ source.
Even in this PR I've caught numerous issues such as incorrect types,
outdated functions, and even some duplicate bindings. Automatically
generating these shifts the focus on making sure the bindings are
correct, and then these will simply fall into place.

*** Please review the new stubfiles for accuracy - Nitpicks are welcome!
***

If desired, I can make a prequel-PR that just manually updates the
stubfiles, and then the machinery can be reviewed seperately.

4 of these stub files are now autogenerated: `dtype`, `driver`,
`engine`, and `profiler`. I've included various typing and doc fixes to
keep them at least as accurate as the previous versions - mainly
parameter names, replacing signatures where necessary, and updating a
few out of sync types. The other generated stubs are excluded for now,
but can (should) be revisited in a followup.

These stub files are expected to be committed - there are `diff_test`s
to make sure that happens. This is to support workflows that might use
these files but otherwise don't want to go through bazel, such as docs
or maybe copybara in the future, if we decide to open source them (which
there's no reason not to, they're already publicly visible)

The new `src` folder in `_core` is so that I can directly `glob` the
`.pyi` files into the `_core` target, which wasn't possible before since
they were in a subpackage. The reason we need this vs a `filegroup` is
that `mypy` doesn't get run unless the `.pyi` files are directly listed
as `srcs`.

For now, we are generating the stubs with 3.12 (i.e. our default python
version). Ideally we generate them with 3.9, as we want to make sure the
stubs are compatible with the lowest version. I still slightly suspect
we're breaking 3.9 with use of `|` for types, but @dukebw thinks it will
be fine - I'm okay with letting this go for now and we can revisit if it
becomes an issue. nanobind-stubgen does some version-specific things,
use of `|` is one of them, but also we want to avoid running on 3.13,
since that will break as `typing.CapsuleType` is only available in 3.13
onwards. So we will need to work something out before then, whether it
be run on a lower version, or patch it somehow.

I originally tried to make this a generic rule in
`modular_nanobind_extension`, but since the SDK extension modules were
all merged into one, and there was a lot of hacky hardcoding needed, I
just moved it directly into SDK. Perhaps in the future it can be cleaned
up and made generic, but we only need it in one place.

Nanobind supports [pattern
files](https://nanobind.readthedocs.io/en/latest/typing.html#pattern-files),
which are pretty useful and will let us simplify some stuff here, as
well as paving the way towards fixing the rest of the stubs (by giving
us rather powerful find-and-replace capabilities). But I want to get
this out the door, since it's pretty complicated as is.

Fixes SDLC-1756

MODULAR_ORIG_COMMIT_REV_ID: b33ccad924a8a1d04f5c9d172f4c56f5965adc67