[Kernels][GPU] Make sure we test the strided load IR (#57092)
Before we were just testing that the code compiles, which is great but
is not the intent. This changes the test to make sure that the IR
contains the llvm.masked.gather intrinsic. Since the generation is
target dependent, we test the llvm code gen rather than the raw asm.

GPU_TEST_ORIG_REV_ID: bb13f67e1ae894fc591266902f472a48d02fdf83