[Kernels] Replace uses of `InlinedFixedVector` with `List` (#57053)
Reduce the use of `InlinedFixedVector` when `List` will easily suffice.
All we were using it for was to reduce allocations when the upper bound
is known. `List` is fine for this in these current uses when applying
the capacity constructor.

This will allow us to remove the old hybrid data structure of
`InlinedFixedVector` soon from the standard library.

Closes: GEX-1155

---------

Co-authored-by: Spenser Bauman <spenser@modular.com>
MOGG_ORIG_REV_ID: 6a1f1ba80c41354254004ac7fd9d34d24daeb0ed