[Support] Move `DType::getWidthInBits` out of line, NFC (#57337)
This moves `DType::getWidthInBits` out of line, since it is a large and
complex method, which is now possible thanks to `getComplex()` not
having to be constexpr.  This also rewrites the implementation of
`getSizeInBytes` to be simpler and less redundant.

MODULAR_ORIG_COMMIT_REV_ID: ffff1d30b2c6c631f707d8b633ca7a016be5220e