[mojo-lang][POP] Move a bunch of string literal stuff to attrs. NFC. (#57346)
A bunch of string literal functionality was implemented as
"non-lowerable mlir operations", which leads to lots of problems
with our model.  This is one step towards fixing this, by
introducing an attribute form of all of these operations.

This is needed to support a future redesign of `StringLiteral`.

MODULAR_ORIG_COMMIT_REV_ID: dcbe2cbfb2820bd6f59560be262816b1d13406e9