[Mojo] Teach the mojo compiler to use ptxas to compile PTX to CUBIN. (#57425)
We observed that using the driver to compile PTX produces different
assembly than using ptxas. This offers a way around that issue and tries
to compile using ptxas if possible. Note that this is a temporary
solution since we'd want to bundle ptxas with the Mojo compiler if this
approach is successful.

Currently using ptxas is only used if the ptxas binary is in the path
and the `MODULAR_USE_PTXAS` env variable is set.

MODULAR_ORIG_COMMIT_REV_ID: 9dd502d20884b6cdba3347218d490c650476e047