[Kernels] Fix UB in sm90 fa2 (#57534)
This applies the optimization from
[Internal link]
to fa2. This optimization is also needed for correctness; the old code's
behavior was undefined, as it was defining the inputs to wgmma
instructions after calling `wgmma_1.arrive()`.

We didn't see the performance problem that #57507 addresses in the fa2
kernel, but I think it's best we get this in for correctness's sake. We
can't rely on the compiler happening to fix invalid code.

NN_ORIG_REV_ID: e3f773df8ceb6c6c1000cb0b54f40922abd6cfc7