[Kernels] Make `fillers` module private (#57573)
It's only used in tests and one example.  None of the public API surface
uses it, so make it a private module.  In the future, we may move it to
the testing directory and update the GPU example.

LAYOUT_ORIG_REV_ID: b185e8d9e29fc5ed2ddfff12d9588237a431dab2