[Mojo] Extend Closure Type to Include Register Passability (#57584)
If a closure is bound to a parameter of a function that is compiled to
an accelerator target, then the captures must be passed by value which
translates to a register passable closure struct since otherwise stack
memory or heap memory will leak into the device function arguments. The
first step in supporting this is adding the ability to express this in
the closure type.

MODULAR_ORIG_COMMIT_REV_ID: db18465702790feb3b3fbb9c497f8e4c52a4f384