[RecordReplay] Add replay functionality (#57315)
This caps off the core functionality of record/replay, adding a
mechanism for replaying a recording of HTTP transactions against an
arbitrary endpoint. This is provided in both library form and with an
executable that can be run standalone. Basic statistics are gathered,
but it's not expected that this will be enough in its current state --
it's expected that future PRs may add more robust reporting
functionality as needed.

Part of AITLIB-239.

MODULAR_ORIG_COMMIT_REV_ID: e2b9c87d82787db3c35d1706450f378d2a1279b7