[Kernels][AMD] extend `vendor_blas` with `hipblaslt` (#57612)
Extend `vendor_blas` to be able to call `hipblaslt` on AMD systems. This
is useful for internal benchmarking/testing and could replace/remove the
existing `rocblas` support. `hipblaslt` supports more data types and is
one of the compute paths used by pytorch. Most of this was based on the
existing `cublaslt` support with bugfixes that should be migrated back
to that version eventually (the transpose flags work as expected).

Closes KERN-1586.

LINALG_ORIG_REV_ID: c2a92e41522fab53776797488674c6a8f303ab55