[CI] [Pipelines] Flake detection for pipeline logit verification (#57714)
We make changes in several areas of the stack to attempt to detect
flakes, retry if it looks like a flake, and if we still fail, detect
this situation and report it as such rather than as a failure.

The portions we change are:

* generate_llm_logits attempts to detect Hugging Face network errors. If
we hit one, we exit with a distinct status code.

* verify_pipelines learns about this new status code to indicate a
flake. Upon encountering one, it will wait a minute and try again. If
that still fails, we note that we think it was caused by a flake and
introduce a new verification verdict to reflect this state.

* Only partly related, but verify_pipelines also is changed to be more
brief about its own failure if generate_llm_logits fails. Normally
generate_llm_logits will have printed its own (long) error message, and
verify_pipelines's traceback can sometimes drown this out; so make
verify_pipelines more concise to highlight the earlier cause.

* The workflow is changed to notice the new "flake" result from
verify_pipelines. If all failures were just flakes, we still report this
to Slack, but indicate the cause as we see it. We also attempt to detect
common conditions related to infra flakes and report these as such as
well. Bisection only runs if we believe we have a genuine failure.

Together, these changes should lead to us encounter flakes that require
human attention less frequently and make it more obvious when one does
occur.

MAX_INTEGRATION_TESTS_REV_ID: 14ece5f3b81affbf84724d39d4f5fdffc7c5c7ba