[Models] Add Phi-3.5-mini-instruct to logits verification pipeline.   (#57853)
Note that the tolerances have to be set very high for this to pass, so
there's an accuracy issue somewhere.
It's still a net positive to get it into the verification pipeline, as
this will detect other types of issues when they happen.

Towards MODELS-429

MAX_INTEGRATION_TESTS_REV_ID: 186160656d0b21ccc27bd719677ac75d501a8a4a