[Mojo] Update Lit Closure Init Representation To Match Kgen Closure Init (#57779)
This patch updates the lit closure representation. There is a kgen
representation for closures but it does not include a representation for
the origin of the struct it represents. We need a way to model not only
the origin of the capture struct, but the origins of the captures in the
event that the value is captured by reference.

MODULAR_ORIG_COMMIT_REV_ID: d1dc771c88961efdf8bbaff303fa2ec8a1071390