[AsnycRT] Disable full buffering when piping stdout if on AMD GPU (#57991)
The default for libc is to do line buffering when writing to stdout,
and full buffering when piping. This was causing the GPU print and CPU
print output to be mixed together, which is why FileCheck tests where
failing because they pipe to a file. This turns off buffering altogether
when initializing an AMD GPU `DeviceContext`, giving expected behavior
without the user having to to think about it.

GPU_TEST_ORIG_REV_ID: 81c952a03d0dcd32a91a9b0753ebb2f637fa787f