[KGEN][CS][NFC] Make CompilationServer to own LLVMServer (#57260)
Stacked PRs:
 * __->__#57260
 * #57259

--- --- ---

### [KGEN][CS][NFC] Make CompilationServer to own LLVMServer

Since compilation server should be self-contained, `LLVMServer`'s
ownership must belong to CompilationServer.
That also solves the problem of keeping `LLVMServer` alive for
in-process compilation server

[SDK] Expose Accelerator.can_access (#57241)

Expose `Device::canAccess` to the Python API.
This is useful in a follow up to stage different ops from the graph API
depending on whether peer access is enabled or not.

An alternative would be to handle this in the
`mo.distributed.allreduce.sum` op lowering.
If that's preferable we can clean it up in a follow up.

MODULAR_ORIG_COMMIT_REV_ID: 7e44e02c81943531d16d01510e9fe0644f0071be