[LSP] Parse time telemetry (#58047)
Adds telemetry for the amount of time, in microseconds, it took to parse
a document. This is reported as an event with duration, size (in bytes)
of parsed code, and the document type (notebook or text).

Fixes MOTO-920.

MODULAR_ORIG_COMMIT_REV_ID: c9fea647e15bd0c80671b3f661a7f257e516ba2a