[KGEN] Expose scale_in and scale_out attributes to mojo wmma functions (#58267)
This exposes the scale_in and scale_out attributes to the mojo wgmma
functions. This allows one to write `D = A @ B` or `D = A @ B + C` or `D
= A @ B - C`. Before we were limited to `D = A @ B + C` which meant that
we had to prefill the C matrix with zeros.

GPU_ORIG_REV_ID: 47f4df83ae4bfc46292224e2100a03fad9195e87