[Kernels][GPU] Add ability to pass StringLiteral to dump paths (#58285)
Previously this would throw an error if you forgot to wrap the string in
`Path`, and the error didn't give any indication of how to fix it, which
has tripped me up a couple of times. Because this is a variant, we can't
rely on `StringLiteral` -> `Path` conversion or use the `PathLike`
trait, so need to add a `StringLiteral` option to the variant. Created
an alias to reuse the same type definition.

GPU_ORIG_REV_ID: c5ee77e82f2d7001c5bc85a40479551be23f4dad