[stdlib] Rename `SIMD.type` parameter to `dtype`
"type" has a specific meaning in programming languages and API design.
In this case, the pattern of `type: DType` as a parameter is not quite
right.  Make it much obvious to callers and readers that the value is
a` DType` (and therefore has its associated methods) rather than a
particular trait value like you may see in collections.

Replace all uses this hits with `.type` with `.dtype`.  In the future,
we could consider using the `element_type` alias we also have for
`SIMD`.

Note: many more places throughout the entire codebase will need to apply
a similar pattern when using `type: DType` in APIs and types.
LINALG_ORIG_REV_ID: e685fc224d5a3ac30327bbe48f4df7b32159fe53