[Kernels][GPU] Add a new version of `wgmma_async` function that accepts StaticTuple Input/output  (#58424)
Our current `wgmma_async` implementation accepts inputs/output type of
SIMD which only works with power of 2 sizes. H100 wgmma instructions can
have shapes which are not power of 2 such as m64n80k16 (check
[here](https://mlir.llvm.org/docs/Dialects/NVVMDialect/#nvvmwgmmamma_async-nvvmwgmmammaasyncop)
for all the supported shapes and types). In this PR we add a new version
that instead of SIMD accepts and returns a StaticTuple which is an
array.

LAYOUT_ORIG_REV_ID: 7e5b7a822e038ebac9eac9de7910c26687f53717