[Kernels] Add Support For Vectorized Advanced Indexing In Some Situations (#58452)
Stacked PRs:
 * __->__#58452
 * #57911

--- --- ---

### [Kernels] Add Support For Vectorized Advanced Indexing In Some
Situations

This PR adds support for vectorized advanced indexing if certain factors
are met.
The correctness of this optimization is based primarily of examining the
store/load which use the `width` parameter in the elementwise lambda.
Such tensors must be contiguous along the innermost dimension.

This was not added before since we did not have a lot of unittests. Now
that we do we can turn this optimization on.

Testing:
 - unittests and integration tests don't break
 - some expanded PT unittests
MOGG_ORIG_REV_ID: fbd57c4d7caf790fdaed7a9d70721ae3f4d33041