[Build] Make all requires-network tests manual (#58607)
This is not a permanent solution. This is to alleviate the spike in
flakes as a part of the current incident. We need to find a way to
stabilize these tests the right way, but for now, disable them.

Part of AITLIB-293.

MODULAR_ORIG_COMMIT_REV_ID: fd9787f74f46d2ab8158c41ecdda11e208d6688d