[stdlib] Add `ParamKVCache` for `Dict`.
New meta-programmed struct to:

`__getitem__[k:K]()->ref[O]V` for each `K` in `Self.List[K]`.

The struct takes a `Keys: List[K]` as a parameter.
Pointers to the values of `Keys` are retrieved during `Self.__init__`.

They are stored in an `InlineArray`, logically mapped to the Keys:
    - `for i in range(len(Keys)): value_pointers[i] = dict[Keys[i]]`

The `get_ref()` method takes a `Key: K` as a parameter.
If that key is not in the `List[K]` struct parameter, an error is raised.
If it is, the `Pointer[V]` mapped to it on `__init__` is returned.

The cache can be updated with the method `update_cache`.

It is then possible to `__getitem__` a `Dict`,
by passing a key previously cached as a parameter.

Pointers are cached on `__init__` once, or with `update_cache()`.

Signed-off-by: rd4com <144297616+rd4com@users.noreply.github.com>