[stdlib] Remove `UInt.lt` from `Int`
This is a footgun left over during the initial bring up of `UInt`: by
supporting `lt` when the rhs is an `Int`, it would ignored the
signedness of the `Int` completely and just compare the two `UInt`
values. This is unsafe and undesirable. Callers should cast explicitly
to avoid surprises in the library.

MODULAR_ORIG_COMMIT_REV_ID: 8184a22ccf7e2518a5fe3772b87be751d8ca304e