[E2E] Tiny micro-optimization for paged kv cache
This is a tiny micro-optimization for the paged kv cache. It first
removes the need to del from the prefetched_blocks and prefetched_cow
dictionaries. Instead, it just pops from the lists. This is a bit more
efficient because it avoids a delete operation. Secondly, it uses
vectorized numpy operations to assign the block indices to the lookup
table. This is more efficient than a for loop.

MODULAR_ORIG_COMMIT_REV_ID: 15e264646bb57a6acd5bef12f6200d545dd7c114