[MAX] Remove MAX visualization support (#58728)
This code is not maintained and is not used anywhere publically, so just
remove it. We can revisit this in the future if needed and do it
properly. In general netron is not a good tool for visualizing large
models, so we should not invest time in maintaining this.

MODULAR_ORIG_COMMIT_REV_ID: daf1775157cdb7c63557a45a8b080de773607359