[stdlib] Upgrade `Slice.__str__` to new out convention `(+capacity=16)`.
Hello, small PR to upgrade,
this also removes a move on the way,
it is quite nice how the two methods work together like this!

Small capacity of 16 (might be enough),
that way, it won't realloc too many times while building the return.

It should be possible to calculate an optimized capacity,
but that seem like a good base line:

```mojo
slice(1, 2, 3)
```

We might need to adjust the capacity, if `None` step is used a lot!

Signed-off-by: rd4com <144297616+rd4com@users.noreply.github.com>