[Runtime] Move CUFile to the runtime module (#58826)
This moves the CUFile implementation to the runtime module. This is
because the next steps of CUFile will require the use of the
DeviceContext and Support cannot depend on DeviceContext. Part of this
is a revert of #58729.

Note that the IO module currently is not complete and has a single
backend, CUFile. In subsequent PRs we will add other backends like Naive
and Mmap.

MODULAR_ORIG_COMMIT_REV_ID: 1e1e93bc3500971365c9495c0abdcda7cd33b90b