[Kernels] Burninate materialized mask codepaths (#58975)
We can just wrap NDBuffers in our MHAMask trait to avoid branching in
our kernel code. Down with branching in our kernel code.

Reintroducing after #58124 was reverted. The main change to this PR is
to retain the logic to fuse the scale_log2e into a single variable if we
aren't using a score_mod. Previously we had removed this manual fusion
to be more generic, but it appears that it could have an impact on
correctness.

KERNEL_BENCHMARKS_ORIG_REV_ID: 6a4828219844056cc7e355a56a50e22e165464f3