[KGEN] Remove POC inst_struct_ref (#59021)
Previously, the operator was used as a way for ParameterEvaluator to
callback into IREvaluator. This PR teaches ParameterEvaluator to do the
callback on not just ParamOperatorAttr, but also TypeConstantRefAttr.
This avoids this wrapper (which decreases handling logic complexity & IR
size).

It does this by introducing a new AttrInterface:
EvaluatableAttrInterface, which represents the kinds of parameters that
need to perform context-aware evaluation in the IREvaluator.

Also had to introduce a concrete version of TypeConstantRefAttr that
doesn't carry param values. It will be substituted in when a
TypeConstantRefAttr is concretized. This way it's clear whether a given
expression is concrete or not just by its construction, and we also
avoid having to deal with the potential of these references carrying
param values post-elaboration.

MODULAR_ORIG_COMMIT_REV_ID: 8dee6c6428dc369c14398cd4b2f85289fec1f263