[Kernels][AMD] set `greedy-reverse-local-assignment=true` (#59064)
Based on the recommendation from AMD in the below LLVM comment, turn on
`greedy-reverse-local-assignment=true` when building AMD device code.
This is needed to fix a problem where the register allocator generates
poor code when using large block sizes that consume many registers.
There is a PR in flight that fixes the problem in a different way that
avoids this hammer and when that lands, this PR could be reverted.

Related to SDLC-1967.

https://github.com/llvm/llvm-project/issues/131954

https://github.com/llvm/llvm-project/issues/131954#issuecomment-2777209408

GPU_ORIG_REV_ID: c8bd5d5ea1577781c3eb00d68d111c9bc62cd62c