[Kernels][GPU] Make the vendor BLAS library work with the global handle (#59085)
This is useful to avoid the overhead of creating a new handle for each
matmul and is used in places where a context cannot be created. We also
do not want to destroy the handle until the program exits.

LINALG_ORIG_REV_ID: 3685e639aa738d6499626b68d59c8e1b5d436b8b