[Kernels][GPU] Simplify the vendor blas calls in tests (#59115)
The vendor blas functions now has a variant which constructs the handle
on its own. This allows us to remove the need to pass the handle as an
argument to the functions. Note that this handle is global and will
always pick the automatic backend, so some tests are left as is because
the ask for more control over the backend and/or lifetime of the handle.

GPU_TEST_ORIG_REV_ID: a0976e7207fc3668f90257ca1dd1eb3db4ddd40f