[Kernels][GPU] Intialize the c buffer if needed for cublas backend (#59137)
When we have an elementwise lambda function, we may not have the output
c buffer allocated. In this case, we need to allocate a temporary buffer
and apply the elementwise lambda function on the result. The code is not
pretty, but I'll clean it up in subsequent PRs.

Thanks to Tracy for finding this bug.

LINALG_ORIG_REV_ID: 31f41d0f271ed5e67b22c7c17f6e0b9302a0b2fe