[Kernels][MOGG][MO] Support Split On GPU (#59026)
Closes GEX-2030

This PR implements the split operator on GPU and enables the use in the
Graph Compiler. The prior implementation only worked on CPU.

Summary:
- Rewrite implementation to support GPU. Note, there are potentially
more efficient implementations that involve parallelizing kernel
launches on GPU. However until we support multi-stream semantics, this
is likely not useful.
- Implement unittest for GPU (there are already lots of unittest for
CPU)
- Plumbing through graph compiler to enable the operator on heterogenous
device.

The unittest was implemented in the graph compiler as this is the most
convenient place to put such a test at this time.

MAX_GRAPH_API_ORIG_REV_ID: acabf34231fb77b4ecb88636905770f727168d78