[MOGGPreElab] Add a pass that verifies kernels as they are written (#58619)
This is a pass that verifies that kernels match our expectations at this
point in the compilation. It is MO-agnostic (except for knowing if
a registered operation is a MO op or a custom op).

The pass uses the `MojoKernelAdaptor` helper. The adaptor itself does
not do any verification. It merely represents the kernel and leaves
verification to this pass and subsequent passes like validating that
a MO operation (custom or not) matches the kernel.

Also moved tests that lived in the Graph API to `MOGGPreElab`.

Fixes GEX-1946

MODULAR_ORIG_COMMIT_REV_ID: 1ae206b57e3fdc9f27b2a2b14bd0def5c4d9821a