[Pipelines] Fix scalar decode
The "Llama whitespace hack" tries to prepend a token before single-token
decodes, and then strips off the decoded value of this dummy token.
However, it uses `np.insert`, which, when the input is a scalar rather
than a rank-1 array, _replaces_ the token instead of modifying it. This
results in empty decodes. Reshape the input to rank-1 if it's rank-0 to
avoid this fate.

MAX_INTEGRATION_TESTS_REV_ID: d025aeeee1f55d927fca1ac47a43dd13ca42e9f4