[KGEN] Separate compilation for GPU kernels with the same backend but different emission options.  (#59090)
- [x] Group GPU offload compilation with the same target info further
into subgroups of emission options so that we can compile kernels with
`nvptx-short-ptr=true` and `nvptx-short-ptr=false` for Nvidia. Note that
this can slow down gpu kernel compilation because the subgroups have to
be compiled in order instead of in parallel because we have to set the
llvm cl options and those are global states.

MODULAR_ORIG_COMMIT_REV_ID: 5f9badd1facfd8c85a33c782b9bcd5017c3cda9c