[mojo-lang] Fix metatype processing hole Stef reported. (#59392)
This enhances `getIfTypeValue()` to work on arbitrary CValues,
including things like SBValues when the metatype is a struct
metatype.  In this case, we don't need to eval the struct value,
because we know it is a singleton. This enables us to look up
static methods (etc) from it.

MODULAR_ORIG_COMMIT_REV_ID: a94e372aa063c293fbe6c3618a52c276b348023e