[SDK] Perform op verification as ops are built (#59494)
Currently, the operation verifiers are run once the graph is finalized
(when Graph.output is called). This defers error messages until long
after the problematic operation is constructed.

It also means that many of the properties of the RMO op that should be
validated by verifers are instead verified in the `inferResultTypes`
method on the backing op.

Closes: GEX-2083
MAX_GRAPH_API_ORIG_REV_ID: 3501da0ad3d70aeec06e9d65a2f6718107bc7fb2