[MO] Support opaque fusion for elementwise producer kernels (#59426)
Adds the necessary support to allow fusion of an elementwise producer
kernel into another kernel that updates an opaque value in-place.

Accomplishes this by adding an overload to foreach allowing users to
pass in an update function to foreach which can embed the logic to
update the opaque value of the consumer kernel.

EXTENSIBILITY_ORIG_REV_ID: 12d775f644e6dfd351e6ce653b24764937785fdb