[Support] Retry on error when trying to take lock files (#59509)
Change #59493 was intending to retry on error when trying to take the
lock file. But the control flow is confusing and the code is instead
returning the callable and effectively ignoring any error. The intent of
the other PR was to retry, so a continue is needed here.

MODULAR_ORIG_COMMIT_REV_ID: f4753c8517efa3af6f600ef1a9b007e391d06e98