[mojo-lang] improve param mangling in two ways (#59603)
Elaborator-time name mangling is in charge of making sure that
instantiation
of functions produces unique names based on the original name of the
function and the parameters it is stamped out with. It has to avoid '@'
symbols
because ELF has a problem with them - it previously replaced them with
'_'
which started causing incorrect conflicts when we mangle two functions,
one like `"foo" with string "_"` and one like `"foo" with string "@"`.
doh.

This changes param mangling in two ways:
1) It removes the leading `@` from function symbols, which cleans up the
   name a little bit and avoids exacerbating #2.
2) This changes the replacement of "@" to be "_at_" instead of "_". This
is
   not a correct fix, but should make the problem less problematic.

MODULAR_ORIG_COMMIT_REV_ID: e76354ce04044be11f0c642fdbb1a5f712b8bd1e