[KGEN][Reland] Don't split private symbol (#58973) (#59505)
**NEW**
After original commit has been merged, nightly testing failed on MacOS
due to new splitting and that LLVM backend generated `private
llvm.global_ctors` in multiple modules. MCLinker renamed these new
functions to avoid conflict, which later resulted that MachineFunction
was not found by AsmPrinter for the new name and compiler failed with
segfault. The taken solution is to change `private` linkage of non-seen
functions to `week` to avoid renaming.

NOTE: this is no longer reproducible on ToT, however this might be a
pure luck and slight change in IR that resulted to change in spltting.
User found a bug when compiler with optlevel > 0 fails with "undefined
symbol". The issue arises from the way we split the module and change
linkage of private global values. Specifically in this case we split a
module

**Old**

```module
:@.memset_pattern = private unnamed_addr constant [4 x float] ...`

define @main() {
  call void @memset_pattern16(ptr %ptr, ptr nonnull @.memset_pattern, i64 44)
}
```

into
```module_0
:@.memset_pattern = weak unnamed_addr constant [4 x float] ...`

; no use of the symbol
```

```module_1
@.memset_pattern = external unnamed_addr constant [4 x float]
define @main() {
  call void @memset_pattern16(ptr %ptr, ptr nonnull @.memset_pattern, i64 44)
}
```

Since `module_1` requires to use GOT and private linkage does not add an
entry to the table, final linkage with library failed.

The solution is to disable splitting for private variables

MOCO-1680
**NEW** MOCO-1751

MODULAR_ORIG_COMMIT_REV_ID: 17294d10688068873dd53af4ef4586f1e4efb211