[SDK] Reapply use new rmo.conv in the graph API (#59634)
Re-enable the use of `rmo.conv`, which was previously reverted due to a
build error in Pixtral. The failure was due to the use of symbolic
dimensions in the Pixtral pipeline. This change loosens the shape
propagation logic so that failing to fold away the assertions at
construction time is not a build failure.

Closes: GEX-1931
MAX_GRAPH_API_ORIG_REV_ID: 5bdbe5a1a534c770aaa56427f00046cebeae190f