[tooling] Migrating trimmed builtin types from ::Int to Int (#59683)
[tooling] Migrating trimmed builtin types from ::Int to Int

This change changes the implementation of `trimBuiltinNamespaces` to
remove the `::` prefix from the trimmed types. Because this is a
behaviour change it requires updating a series of tests that are
validating the generate output for mojo docs and mlir code.

MODULAR_ORIG_COMMIT_REV_ID: 73d537ebe0fed4aedb3de4f5a64be9d307afe73b