[MOGG][Extensibility] Improve logic to infer IOSpec in MoToKGENLowering (#59636)
In MoTOKGEN, we don't have directly access to the IOSpec parameter of the kernel operands.
But we still need it for lowering, so we have a complex logic to recompute it from several decorators / attrs.
This PR adds a new kernel decorator storing the IOSpec values directly to fetch them later.
I also updated the logic to fetch it using MojoKernelAdaptor which is now the correct way (instead of manually looking at the kernel attributes).

This is needed to support a new type of output lambda, it makes it a lot easier to propagate the IOSpec correctly through MoTOMOGG.

MODULAR_ORIG_COMMIT_REV_ID: 3aac52d2b00083d942e74eb329e54414b1ed281d