[ModularFramework] Rename libmof -> libmax (#59780)
Towards unifying into a single shared library. libmof has started to
take the place as this, so let's make it a north star and name it
something more obvious. There's going to be lots of cleanup after this,
but this is just a rename.

MODULAR_ORIG_COMMIT_REV_ID: 785dd6f8a93fa68bc6f6264a4a27e672c741c693