[SDK] Rename `Tensor.onHost` -> `Pinned`, add `pinned` accessor (#59621)
Stacked PRs:
 * #59622
 * __->__#59621

--- --- ---

### [SDK] Rename `Tensor.onHost` -> `Pinned`, add `pinned` accessor

`onHost` is a confusing name for a user because we have a separate
notion of `is_host` already, which is `Tensor.device == CPU()`.

My understanding is that this is the same notion as PyTorch's
`pin_memory`, aka page-locked memory, which allows more efficient
transfer to device memory.

This PR renames these arguments specifically in the SDK, leaving the
underlying Driver implementation alone, and adds a `.pinned` accessor to
allow users to check if the memory is pinned.

MAX_TESTS_TESTS_ORIG_REV_ID: 20895b7327c887f630e7a2e5cb41c95501147fc5