[kernels] Demo overlap matmul and allreduce.  (#59845)
Add PDL to allreduce and matmul and a test to show how to overlap them.
The test itself doesn't achieve much overlap because matmul has a
`wait_on_dependent_grid` call serializes itself with allreduce. If we
remove it, which is just safe for the test, we achieve the overlap shown
below with ~20% speedup.

![overlap-matmul-allreduce](https://github.com/user-attachments/assets/09c6f552-030a-4d7a-9040-f52791e2d0c0)

To enable a solution E2E, we need to parameterize pdl behavior by layer
and generalize the partition (test only partitions in M).

GPU_TEST_ORIG_REV_ID: 4b7b7ab4959224094af79740913baf1738c17cc2