[mojo-stdlib] Improve encapsulation and safety of String.
This patch is a step forward to improving the encapsulation and
design of String.  It uses the new `unsafe_uninit_length` ctor to reduce
cases that form a `List` and then transfer it into the `String`.

This also improves "origin correctness" in `StringSlice` a bit.

MODULAR_ORIG_COMMIT_REV_ID: aa3c75cd4d8e6509344c7b9acd33980018259300