[KVCache] Refactor block copying logic into a new class
Now there is a `BlockCopyEngine` that is responsible for enqueuing the
memcpy ops.

Previously we staged copy ops and only enqueued them when the req was
successfully scheduled.

This means that if we stage a D2D COW memcpy and there is not enough
blocks available for that req, we discarded the D2D COW memcpy.

It turns out this is largely unnecessary since scheduling a single extra
block copy is not a big deal.

Now the block manager directly enqueues the memcpy instead of staging
them until later, eliminating the need for such things housekeeping
structs like `req_to_block_copy_ops`.

We also provide two different strategies for offboarding blocks from the
device:
- EAGER: Immediately copy the gpu block to host memory once it is
committed.
- LAZY: Only copy the block to host memory when it is evicted from the
gpu cache.

Generally the EAGER strategy is better assuming that we can schedule on
a new stream.

SERVE_ORIG_REV_ID: 60f28198a8b6887f835002ed444ff0496f8a251e