[Stdlib] Add exp trait and overload for exp function
This allows us to be able to use math.exp on any type that conforms to
the _Expable trait. This is going to be used in subsequent commits to
allow us to use math.exp on complex numbers.

NN_ORIG_REV_ID: 986e1f61dcc8c2670705b685aad8fe7691d076f3