[Kernels] Remove symlink to `test_vec_add.mojo` file (#59961)
The `test_vec_add.mojo` in `Kernels/test-gpu-debugging` is a symlink to
the one in the examples file.  However, the RUN directives etc make it
so it's only useful in the debugging test.

Avoid future confusion by removing the symlink and just moving the test
from `Kernels/test/gpu/examples` to `Kernels/test-gpu-debugging`.

Note the test is disabled anyways, but it got in the way with other work
I was doing with licenses due to us not following symlinks.

GPU_TEST_ORIG_REV_ID: f0dc4034724eb1bc6918ab64757c7a5760dfa7c9