[mojo-stdlib] Change `sStringSlice[a:b]` to return `StringSlice`
This changes slicing of a StringSlice to return a StringSlice
instead of forming a copy of the String data.  This makes it more
comparable to Span. This fixes MSTDL-1399

This was coauthored by https://github.com/martinvuyk in
https://github.com/modular/max/pull/4319

MODULAR_ORIG_COMMIT_REV_ID: 4413655fd71edda4638e8fa462a89789895e1cf7