[Kernels][GPU] Cache the P2P check result (#60027)
The P2P check is not necessary to be called on every allreduce call. We
can cache the result in a global variable, since the result is invariant
for a given device pair and is constant for the lifetime of the program
(since we do not expect one to just swap out the motherboard mid
execution).

GPU_ORIG_REV_ID: 27795ff02d62f9259d8bb087024833fa3b235893