[LSP] Fix debugger auto-attach (#59920)
This was a nightmare to track down, but essentially attaching a debugger
to the LSP was broken because we were not suppressing output of the
child process that actually starts the debug session. This wrote invalid
output to the language server's stdout stream, which broke the VS Code
LSP proxy, which caused the LSP to seemingly hang.

Fixes MOTO-1047.

MODULAR_ORIG_COMMIT_REV_ID: 8c42378d316e92565846f43c6371308b83f0276f