[KGEN] Promote by value struct to by reference for kernels (#59536)
The patch promotes argument of a strut type with pointers to argument of
a pointer to a struct type with a special `by_val` argument convention,
which is later lowered to either `byval` or `byref` attribute in LLVM
IR. This is needed to allow to emit better code on NVPTX or AMDGPU that
can propagate addrspace properly to pointer within a struct.

Related MOCO-1812.

Co-authored-by: Tracy <tracysh@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 028fd19ce42532ae6e2812f966fda0608e37c244