[M] Enable properties storage for non-discardable attributes (#60159)
This option was disabling properties storage for the inherent attributes
of operations. This option can make working with inherent attributes
more efficient and avoids accidentally polluting the MLIRContext with
attributes which live forever.

Additionally, the usePropertiesForAttributes option is slated to be
deprecated at some point, so it is good to be ahead of that schedule.

MODULAR_ORIG_COMMIT_REV_ID: 05922c70992c23c6b9491c9481bd8533fbb68d5d