[Kernels][AMD] fixes for rocblas/hipblaslt vendor blas paths (#60189)
Some standalone changes that I needed for trying out
`MODULE_USE_VENDOR_BLAS` inside benchmarks:
1. Fix the rocblas backend to set the stream handle as is done for
cublas.
2. Fix the weird hipStream_t code that I had introduced before to be
more rational.
3. Fix a test that didn't work with vendor BLAS turned on. There were
still a few other tests that aren't working right but that's not a
problem for today.

It's possible to run AMD benchmarks with ROCBLAS as the default backend
with no problem.

Using HIPBLASLT is a little trickier because pytorch loads its own copy
of the shared library and this interferes with our attempt to load the
library from /opt/rocm: there is code in libhipblaslt.so to go find its
DLL path and it finds the wrong version and uses incorrect tensile data
files. It's possible to hack our loading code to load just
`libhipblaslt.so` and bind to the version already loaded. That was good
enough for the data collection I was trying to do.

GPU_TEST_ORIG_REV_ID: 8d3dabb1efa3777d29fd6d629494317923d77639