[KGEN] Rename some of the functions in GPU module if they are both callee and kernel entry function.  (#60387)
Mojo GPU kernel compilation bundles all kernels in the same MLIR module
for one target. When a kernel entry function is also a callee for
another kernel in the same module, we need to rename the kernel entry
function wrapper name so that we don't have symbol name collision for
two functions that are now having different bodies:
- The kernel wrapper function would call the generated kernel
`asm_stub`. This only happens when the kernel entry function is a
parametric function.
- The original function as a callee should stay the same and not being
exported.

This doesn't always happen since the generator can be instantiated with
different parameters without symbol collision, but we need to address
the cases when that happens.

This PR addresses the issue by renaming the callee with a postfix
`_callee` so that we can keep the two copies of the function body
without confusing kernel entry function with a different name.

MODULAR_ORIG_COMMIT_REV_ID: e63d1cdc767038893f692f4297ae1dfb1827d436