[KGEN] Fixed flaky behavior of parallel call to `mlir::verify` (#60480)
`mlir::verify` relies on `DiagnosticEngine` to be set in order to report
an error.
The `DiagnosticEngine` can be set using `ScopedDiagnosticHandler` class,
however it's not thread-safe which result to a race condition within
MLIRContext of handlers to report error into.

MLIR does have skeleton for parallel diagnostic handler, but it does not
provide necessary API and seems to be untested. We may need to consider
to extend it.

MODULAR_ORIG_COMMIT_REV_ID: e80d1e6743902035f15cd4370614c9f666bc8ab3