[Kernels] implement buffer slicing for `KVCache.blocks` (#60551)
Changed the two major `KVCache` implementations to use buffer slicing
for the per-layer `KVCache` buffers.

In the existing scheme, the `KVCacheCollection` would create a `KVCache`
that had a copy of the 6D tensor shape/strides plus the `kv_idx` and
`layer_idx` to use for partially indexing that 6D tensor.

This indexing is unnecessary inside the kernel: in the new scheme, the
`KVCacheCollection` computes the address of the portion of the buffer
for `kv_idx` and `layer_idx` and then maps that with a 4D tensor that
the `KVCache` object can use. The kernel has less indexing to deal.

After this PR, there is still one stride for the kv paged buffers that
is not statically compiled into the kernel: this stride needs
`num_layers` to be passed down through `KVCacheStaticParams` too.

GPU_TEST_ORIG_REV_ID: 123b60956ebc8da39c62374d4bfb7191e46280db