[LSP] Add a debug notification to emit parsed IR (#60586)
This allows sending a `mojo/emitParsedIR` method to the language server
to ask the language server to dump the output of the parser to a
`lsp-parsed.mlir` file.

This method is disabled in production builds of the language server. The
corresponding VS Code command will only appear in development runs of
the extension (when running from source code with a debugger attached).

MODULAR_ORIG_COMMIT_REV_ID: 86c1ed78560fa4256403b23fe313b341b25c5c17