[MOGG-PreElab] Annotate kernels that use 'out_func' overload of foreach (#60414)
Stacked PRs:
 * #60415
 * __->__#60414

--- --- ---

### [MOGG-PreElab] Annotate kernels that use 'out_func' overload of
foreach

Adds a bit of logic to the MOGG pre-elab to annotate kernels when using
the overload of foreach that takes an 'out_func'. Annotates the
kernel with the name of the kgen param decl region corresponding to the
out_func lambda.

Neccessary when writing a kernel that does an in-place update to a
mo.buffer value (that we want to be eligible for epilogue fusion) since
we don't have a standard output lambda in such cases.

MODULAR_ORIG_COMMIT_REV_ID: 7497355836c6bafa3fc5d76ecb92bc5e53047041