[mojo] Fix crash when resolving calls to parametric trait methods in the presence of composition (#60603)
When checking conformance with a trait that defines a parametric method,
the parser can
crash because the self type on the method is never prechecked, while the
type parameter is.
This is fixed by upcasting to a parent trait type, which should always
be known for the self
type of trait methods.

MODULAR_ORIG_COMMIT_REV_ID: 77d2e2d862a98f23528c50d054cee4dba4bb0d88