[SDK] Create test fixture to initialize the MLIR context and KernelLibrary (#60564)
The Graph constructor currently creates a new MLIRContext and
KernelLibrary when they are not supplied (and it is impossible to supply
a KernelLibrary directly to a Graph). This causes each Hypothesis test
to create a new MLIRContext and KernelLibrary for every execution of the
tested function.

This becomes expensive when testing custom ops, as the KernelLibrary has
to be fully loaded to validate the custom operation. This results in
each test iteration taking ~0.4s on my machine and Hypothesis typically
runs 100 iterations.

This change creates a graph_builder test fixture which allows all test
runs in a module to share the same MLIRContext and KernelLibrary, paying
the large initialization cost only once.

MAX_GRAPH_API_ORIG_REV_ID: 1e6671517b492161df10acda3a59ce958883dfdb