[bazel] Fix tests discovering system python (#60886)
When using the python C API, it's possible that Py_Initialize auto
discovers /usr/bin/python3, if no other python3 is in the PATH. For the
`modular_python_binding_library_test` we were attempting to steer it
towards the bazel python with various environment variables, but it
turns out `__PYVENV_LAUNCHER__` only works for this for python 3.10+, so
tests would pickup system python if you tested with python 3.9.

With this change we break global python discovery by setting PYTHONHOME
to /dev/null, and then we override that in the few places that need it.

MODULAR_ORIG_COMMIT_REV_ID: fca8cd33c2835b585533ccfc40b457571cb70983