[Stdlib] Add an MLIR type method to DType and use it in WGMMA ops
This adds a `__mlir_type` method to `DType` to get the MLIR type of the
DType. It is used in the WGMMA ops to get the correct MLIR type for the
operands and results. Using this we can simplify the
`dtype_to_nvvm_type` function in the WGMMA ops as well as generalize
wgmma to support more data types.

MODULAR_ORIG_COMMIT_REV_ID: 5506f165a5e7a4792a3ab44a6c175ae6e2113a70