[External] [stdlib] Support for RTX 2060 (#61058)
[External] [stdlib] Support for RTX 2060

<!--
Thanks for submitting a pull request,
your contribution is really appreciated!

If possible, add a link to the issue you are
trying to solve in the pull request description.

If your pull request is big (> 100 lines), consider splitting it
into multiple pull requests as it may accelerate the review process.
-->
This PR adds support for the Nvidia RTX 2060, which I have. I've updated
the file with the info. I'm a bit unsure about the section on MLIR
attributes, so I’d appreciate any advice or clarification on that part.
For the benchmarks that don’t rely on libraries marked with `_internal`,
I was able to run them without any issues. So it works, but I haven't
compared the performance with CUDA yet.

The original discussion about Turing architecture support can be found
here: https://forum.modular.com/t/support-for-turing-architecture/1304

---------

Co-authored-by: Slava Chaunin <67190162+Ssslakter@users.noreply.github.com>
Co-authored-by: Joe Loser <joe@modular.com>
Closes modular/modular#4480
MODULAR_ORIG_COMMIT_REV_ID: c17c8212bcb88721e540dd49ecce00aabdac3321