[KGEN] Split out contextual evaluation policy from ParameterEvaluator (NFC) (#61120)
In preparation for context-aware parameter evaluation / simplification
outside of the elaborator (necessary for simplifying GetWitnessAttr).

Today, a subclass of ParameterEvaluator needs to handle two kinds of
state:
- Parameter bindings: This is state specific to a ParameterEvaluator
instance.
  - This is local to a parameter context.
  - This determines the policy for replacing parameter reference attrs.
- Global context: This is the contextual info for evaluating
context-aware parameters.
- This is the same for all ParameterEvaluators created for a given
module.
  - This determines the policy for replacing ContextuallyEvaluatedAttrs.

For evaluating a context-aware parameter like GetWitnessAttr, we need
the invoker of ParameterEvaluator to provide the global context, which
only needs to be created once (for a given stage of compilation). This
needs to work even inside "3rd-party" helper functions like
getSpecializedGenerator. Instead of passing a concrete instance of a
ParameterEvaluator to these helper functions, it's cleaner to instead
pass the state for evaluating context-aware parameters. This decoupling
of state (and thus, policy too) is better in two ways:
- If the 3rd-party helper function needs multiple ParameterEvaluators,
it can now create them as they'd like and just inject the policy into
it.
- The global context only needs to be created once and can be saved if
it ends up being used in many places. It is essentially just a "policy
server" that can handle evaluating ContextuallyEvaluatedAttrs.

This change simply breaks out the virtual `evaluateExpression` of
ParameterEvaluator into a separate class that is dedicated to it. This
makes ParameterEvaluator no longer have virtual functions. The vtable is
instead carried by the policy class. It does not change the performance
characteristics when a policy class is provided, but will be ever so
slightly faster when the policy is not provided (since that wasn't an
option before).

MODULAR_ORIG_COMMIT_REV_ID: 0194694feecef4f355ea9a2ea37f026cd55fdb5e