[KGEN] SymTabEvaluationContext handle GetWitnessAttr (#61202)
Populate the base `SymTabEvaluationContext` with logic for evaluating
`GetWitnessAttr`.

Sub-class SymTabEvaluationContext with a LIT-level version that handles
evaluating parameters in a LIT context. It delegates anything it fails
at to the base KGEN version.

This sub-class is used in VerifyParameters so that any verifications
that rely on ParameterEvaluator (getSpecializedGenerator) can operate on
ContextuallyEvaluatedAttrs. Eventually we can consider a flag on
VerifyParameters to skip the LIT-version evaluators once LowerLIT
happens.

MODULAR_ORIG_COMMIT_REV_ID: 6243d75a17b4183e17dc8f8681d883763d3720ed