[KGEN] Elaborator put ImplNode directly in ParamNode (#61423)
Since we now have a 1:1 relation between ParamNode and ImplNode, and the
[previous [Internal link] made it
so that both are always created at the same time, we can now avoid half
the mallocs by just putting the ImplNode directly in the ParamNode. This
resulted in a teeny tiny speedup in the elaborator pass 🤣 nothing
groundbreaking, but why not.

Also removes `ImplNode::baseName` because no one was using it (except in
a debug print, which was improved too).

MODULAR_ORIG_COMMIT_REV_ID: 22248677e961ed474e587b8331bd3b1d2d887322