[mojo-parser] More explicitly handle to-trait conversions (#61522)
The parser should do an additional check when converting to a trait that
what's being converted has the right "kind" in the first place. Namely,
it should only be a struct-metatype or another trait. Only then can the
`doesNominalTypeConformTo` check do its job. Previously it was letting
in other kinds of types in and relying on error-checking when emitting
vtable to catch it, which is way too late and results in bad error
messages.

MODULAR_ORIG_COMMIT_REV_ID: 7795982cbbc1e60a65fdd2235c366d6a4c12c687