[mojo-parser] Fix ConformanceOp resolution failure checking (#61528)
Actually need to mark the decl as erroneous when resolution of
ConformanceOp decl fails. And when checking multiple conformances, allow
all to go through before returning false. This does not make a
difference currently because when emitting the actual vtable we'll do
another set of checks and another set of errors will be emitted, but
will be required when we remove vtable emission in
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 8ae40252cb82075115504e56e44ae0628dd43e03