[Mojo] Unpack `kgen.pack` Types (#59775)
This refactors the lower arg pass to unpack the kgen packs. I refactored
the logic so the transformation occurs in place instead of saving
mapping information and applying it after the transform. The reason I
did this was because it simplifies the information necessary to store
the unpack operation.

MODULAR_ORIG_COMMIT_REV_ID: aa519882995a29b5eb5f26313f7b4267db216882