[mojo-lang] Rework destructor emission to be smarter/simpler. (#61610)
All (implicitly destructible) structs get a destructor because they
have to conform to AnyType.  Many of them are trivial though, not
having to do anything in their destructor. This predicate doesn't
have much to do with RP-Trivial, it is a simple compositional
property based on whether the destructor was declared with custom
logic and whether the fields had non-trivial destructors.

Keep track of this in the IR, and use that information to elide
calls to destructors in CheckLifetimes, a micro optimization.

MODULAR_ORIG_COMMIT_REV_ID: 444b4d25aaab07c947e8069abc9c29797482cf99