[CI] Remove use of `--remote-header` in new Bazel job
The `BUILDBUDDY_TOKEN` secret is only available to Modular employees part of the
new org we set up for public CI.  However, this means when we always specify
`--remote_header=$BUILDBUDDY_TOKEN`, it will expand to an empty value for
non-Modular folks.  This creates a parsing error and the job never runs.

Workaround this for now by just removing the `remote_header`.  There's other
things we need to figure out if we explore remote builds and caching for
non-Modular contributors.  This unblocks the job from being experimented with in
the non-Modular contributor case.

Signed-off-by: Joe Loser <joeloser@fastmail.com>