[mojo-lang] Move ExplicitlyCopyable synth to be lazy. (#61800)
This moves the body synth logic for ExplicitlyCopyable to be lazy
and on demand, which means that we don't need the fields to be
resolved when the signature is created.

While here, improve an error message and rearrange some code.

MODULAR_ORIG_COMMIT_REV_ID: e70535a70419e4a26444c62912113b51b514490e