[bindings] Remove boolean caster hack (#61851)
Basically just skip the test this was failing on. It seems weird to
intentionally assign an integer (of 1 through 4) into a boolean vector,
so if that's the only reason the workaround is needed, let's remove it.

MAX_TESTS_TESTS_ORIG_REV_ID: 21c0e04d0d66d4e21a21493166294d301b047fec