[stdlib] feat: Add `DLHandle(flags)` constructor, for getting a handle to all global symbols
This wraps the behavior of `dlopen(nullptr, flags)`, which is useful
when trying to `dlsym(..)` symbols that are known to be provided by the
current process main executable or other previously loaded dynamic
libraries.

This also refactors `CPython()`'s constructor to use the new `DLHandle`
constructor in the case that the `MOJO_PYTHON_LIBRARY` environment
variable is set to `""`. That special value indicates that `CPython()`
should look for symbols in the main executable. This avoids a narrow
problem in cases where:

1. Mojo code has been compiled into a Python extension module (dynamic
library) that is being loaded by Python.
2. The main `python` executable statically links the CPython
implementation
- This is typically not the case on macOS, where `python` is typically a
small executable that dynamically links to a Python implementation
dynamic library.
3. Attempting to `dlopen(..)` the Python `sys.exectuable` directly will
fail because the main process image is a PIE executable
(position-independent code), and that is no longer supported on Linux.

MODULAR_ORIG_COMMIT_REV_ID: 523445fbb6c712c9ca7f32c35b86e2d7e71c9570