[KGEN] GeneratorAttr (#61717)
A GeneratorAttr is the immediate / anonymous form of a KGEN generator.
It is the building block for parametric aliases and the missing piece in
the KGEN generator system.

This also introduces ParameterScopeAttrInterface, which is the
counterpart to ParameterScopeTypeInterface. This allows attributes to
declare parameter scopes and for nested attributes to refer back to
them.

MAX_PYTHON_ORIG_REV_ID: 8c77ebf095df404b602f9a358a92e9b3bfd6e8c4