[SDK] Use a more Triton-like API for interacting with custom kernels (#61878)
The current workflow takes a value-semantic approach to how custom
operations are exposed to PyTorch. This structure requires users to
explicitly supply a `FakeTensor` function before use. Unlike normal
torch operators, our custom operations _always_ require the registered
FakeTensor function.

This change is a prototype for a more Triton-like API for the registered
kernels. The primary change is that the compiled kernels are interacted
with in a DPS manner and users wrap the kernel in a function which both
allocates the output and provides shape information to the compiler.

MAX_TESTS_TESTS_ORIG_REV_ID: 2fa8a5d8b2cc738fe927ac3f2c4846f7af8e0100