[KGEN] InlineParametric shared NameUniquer (#61699)
Fixes MOCO-1926.

Make InlineParametric share a NameUniquer for the inlining of all calls
inside one function. Profiling shows that constructing a NameUniquer
absolutely dominates the time spent in this pass (> 98%) on this test
case.

This change makes it so that the NameUniquer is kept up to date by each
call that's inlined. This has a minor change in the names that this pass
outputs, but IR remains alpha equivalent. There is no observable impact
on the compile time of more typical (non-pathological) test cases, but
does make this particular test case compile in a more reasonable time
frame.

MODULAR_ORIG_COMMIT_REV_ID: decf7367a31bd4aba360319ee3b949a9c7453da2