[python-bindings] docs: Move `basic-pybind` to KGEN + polish bindings tests README.md (#61943)
This is in preparation for making the tests directory public in the near
future:

    //open-source/max/mojo/integration-test/python-extension-modules/

The basic-pybind test covers the very incomplete `--gen-py`
functionality for compiler-driven bindings, which are not ready for use
and not actively being developed.

To prevent user confusion about how the bindings work, we'll keep this
test internal for now, so that the only publicly visible way of writing
bindings is using the "manual" bindings APIs.

Additionally, make some minor cleanups to the README.md in the tests
directory.

MODULAR_ORIG_COMMIT_REV_ID: 830ee9e087474c6c67fc521a35b7c4398b02b686