[mojo-stdlib] Add a microoptimization to Tuple, NFC
This adds a microoptimization to Tuple to avoid doing work for an
empty tuple.  This shouldn't matter after optimization, but reduces
the amount of IR and gruff that gets generated in the trivial case
that the optimizer has to inline and remove.

MODULAR_ORIG_COMMIT_REV_ID: 926ff413f7f957bc5b40bed0d9d12c667fa21ed7