[mojo-lang] Implement initial support for list literals.
This teaches the compiler to emit list literals by passing the values in
the literal to a variadic initializer that takes the values as well as a
`__list_literal__: ()` keyword argument.

Currently this only works when there is a complete inferred type, there
is no defaulting to List yet.

MODULAR_ORIG_COMMIT_REV_ID: c424ab23200a0dcbea0e7c1beb785134fc036499