[mojo-stdlib] Default uninferrable collection types to `List`.
This patch teaches the compiler to handle list literal that have
elements
but that lack a contextual type.  The algorithm merges the types of
all of the elements (using implicit conversions or `__merge_with__`) to
form a consistent element type, and use that as the element type of
`List`.

This is nice because it properly handles conversions, e.g. turning `[1,
1.0]`
into a `List[Float64]` and handling things like `[ptr1, ptr2, ptr3]` by
using `Pointer.__merge_with__` to find a common origin for the pointers
in
the list.

MODULAR_ORIG_COMMIT_REV_ID: 06b5b3e33ecc91b3cb499da48f674418a7c3861d