[SDK] Raise on unsupported multi-GPU codepaths
Raise an error when using stacked MLP or GPTQ weights with multi-GPU.
This is unsupported and raising an error is preferable to silently
falling back to single-device for predictability.

As a drive by, plumb in the `Float8Config` to `Linear` and
`AttentionWithRope`.

MODULAR_ORIG_COMMIT_REV_ID: 6221721c6a8fe55769acd1af6989b511a8640b59