[mojo-lang] Remove the "dict subscript" extension (#62038)
This was an old extension we added early on in Mojo to support
direct construction of types.  The logic for that got removed
quite awhile ago, and the syntax isn't needed, so remove it to
simplify the code.

MODULAR_ORIG_COMMIT_REV_ID: a4343a746f8a76de4f1203bdb82d1fe58bb0496e