[mojo-lang] Implement support for set literals in non-inferred ctx.
This teaches Mojo that initializer lists without an inferred type and
without keyword arguments are set literals, for example:

`var set = {1, 2.0, 3}` is a `Set[Float64]`.

Note that this does not handle the case when the contextual type
is known correctly, e.g. `var set: Set[Int] = {1, 2}` is not handled
correctly.  That will be addressed in a follow-on.

MODULAR_ORIG_COMMIT_REV_ID: 19bb35cbed22806e56d8f4cd0e0b134710885d68