[python-bindings] fix: Fix illegal `__del__()` of Python Mojo objects that errored on init
Due to a subtle distinction I missed between Python's `tp_new` and
`tp_init`, we were
incorrectly running `__del__()` methods on bound Mojo types that had
merely been
allocated, but not intialized with a valid Mojo value.

By default, Python zero-initializes all new allocated objects, which
masked this issue,
since the previous `String` implementation was valid when all fields are
zeroed. But
the new optimized `String` layout is not valid when zero initialized,
which exposed
this bug.

The prior code would lead to undefined behavior due to the following
sequence of events:

1. User initiates construction of new Mojo object from within Python
code

2. Python calls type `tp_new` callback function to allocate space for
the new object (whose
default implementation that we previously relied on); memory is
allocated and filled with zeros.

3. Python calls type `tp_init` callback function, which raises an
exception **without initializing
    an instance of the Mojo type in the allocated memory.**

5. Python calls type `tp_dealloc` callback, which we define as **calling
`__del__()`** and
    freeing the object's memory.

The problem is the mismatch between (3) and (4), where we
unconditionally always
call `__del__()`, even if the object was never fully initialized in the
first place.

The fix is to move initialization logic earlier, into `tp_new`, to
ensure we are not assuming
that all Mojo types can be validly zero-initialized.

This also fixes our `tp_dealloc` wrapper to actually free the object
pointer, which we've been remiss in doing.

PYTHON_BINDINGS_ORIG_REV_ID: 803be5e18fa9eb329ad5c7453c92067a0be06687