[KGEN][AMD] scalarize rounding bias in fp32->bf16 cast  (#62092)
Change the constraint for the constant "roundedBias" to be a scalar
register instead of a vector register. That is what you'd see in the
default generated code for fp32->bf16, but I had assumed all operands
needed to be vector registers in the original mojo implementation.

MODULAR_ORIG_COMMIT_REV_ID: 01a5223ac56fb1c4e038af533d08798021bc290c