[SDK] Use a new MLIR context for every Graph build & compile (#61946)
Re-using a single MLIR context tied to the CustomOpLibrary causes
significant overhead when multiple graphs must be build/compiled.
For our existing unittests, this adds ~10 seconds of total runtime to
each test point after the first one (when the compiled object is not
cached).

I suspect this is due to the accumulated state on the context,
particularly the registration of a large number of canonicalization
patterns for dialects that only appear very later in the pipeline.

MAX_PYTHON_ORIG_REV_ID: 865860de8d96fef174ca2adac0a1fd8d11314756