[MOGG] Improve error messages for custom op params (#61932)
If custom op parameters have the wrong type, provide a useful error
message that indicates the name of the custom op, the name of the
parameter, the value passed to the parameter, and the type that value
should be convertible to. This error message is provided via the Python
SDK as a `ValueError`.

This is not the final form of these error messages, but it does give us
an incremental improvement over the assertions we had before.
Specifically, these errors happen at model compilation time (e.g. in
`session.load`) instead of a graph build time. Once MAXPLAT-161 is
complete, then we should move these errors out of `ParameterMapper`
(since it won't exist anymore) and put them into `MojoLibraryAnalysis`
so they can occur earlier, at graph build time. But at least they happen
before `JitCompileKernels` in model compilation, so they are still early
enough to be valuable to the user.

The new tests added are a little slow - they take about 5 seconds each
on a macOS machine. This is likely because we need to initialize a new
`InferenceSession` for each test, which is slow. I don't see a better
way to isolate the graphs though. If we reuse the same session, the
errors from one bleed into the other, and make validation difficult.

Fixes GEX-1768

MODULAR_ORIG_COMMIT_REV_ID: e64032d4e0b9ee83f47b32170bef12d4edfef6e0