[CI] Fix skipping autotune tests in Bazel
https://github.com/modular/modular/commit/997204763fd77afc64eac0d82e716a3766711a28
broke the open source Bazel build due to directory moves.  However, note
`autotune_tests` is defined with `modular_py_test` which is a no-op in
the current OSS Bazel build.  As such, the Bazel target doesn't exist
anymore.  So drop it form the GitHub workflow as you can't explicit skip
a target that doesn't exist: that results in a hard error in Bazel that
we see now.

Signed-off-by: Joe Loser <joeloser@fastmail.com>