[SDK] Fix pytorch custom ops
Fixes a regression in the nightly with pytorch custom ops, as well as
fixing the unit tests to catch this issue in the future.

The details are relatively boring, we happened to have a test fixture
providing a global mlir context, which then masked the failure when run
in a normal environment, in conjunction with a change which removed the
active context for computing the key.

MAX_PYTHON_ORIG_REV_ID: be372c958d6e4db591f2a11aed578bb6842d11f9