[KGEN] Body resolve all bytecode decls (#62187)
This patch removes granularity from
`SharedState::resolveDeclFromBytecode`. It does not appear to be safe to
only signature-resolve a decl that we have loaded from bytecode, because
doing so may violate IR invariants.

This fixes MOTO-1076, where we encounter a `FileModuleOp` in the
standard library package that is never loaded beyond having its
signature resolved. This never materializes its body from bytecode,
meaning it stays empty. This later violates an LLVM invariant that the
`FileModuleOp` have a single `Block` in its body region (because it is a
`SymbolTable`), causing us to crash.

I do not know for sure if this is isolated to just `FileModuleOp`, and
without a systemic review of our op structures it seems unlikely that it
is safe to not materialize IR in this case.

This change should have no effect outside of the Mojo language server,
because other users of the parser ask for a recursive body resolution of
all included decls.

This also fixes MOTO-1063.

MODULAR_ORIG_COMMIT_REV_ID: 9f8d21f3867e87346cea11815959d2dd4a45ba8f