[LSP] Reject symbols with out-of-bounds ranges (#62262)
The parser started to interact with the language server oddly: we would
create a symbol with a range just slightly out of bounds of the main
document, which would then be rejected in `SourceMgr` because we
wouldn't be able to find the buffer that contained it.

We fix this by rejecting symbols whose start or end are not contained in
the main document.

Fixes MOTO-1041.

MODULAR_ORIG_COMMIT_REV_ID: 9ed5681b1b35d6447f4aff028e406c05454452a2